chore: Update comment controller to include comment update functionality
This commit adds the functionality to update a comment in the comment controller. It allows the user to update the content of a comment by providing the commentId and the new content in the request body. The comment is then updated in the database and a success response is returned.